+++ src/Device/deviceInfo/deviceInfo.ts
@@ -28,7 +28,9 @@
    through CSS, e.g. to change the colours.
 */
 
-import type { Device } from '../deviceSlice';
+import type { Device as NrfdlDevice } from '@nordicsemiconductor/nrf-device-lib-js';
+
+import type { Device as WrappedDevice } from '../deviceSlice';
 
 import nrf51logo from '!!@svgr!./nRF51-Series-logo.svg';
 import nrf52logo from '!!@svgr!./nRF52-Series-logo.svg';
@@ -38,6 +40,8 @@ import nrf91logo from '!!@svgr!./nRF91-Series-logo.svg';
 import ppkLogo from '!!@svgr!./ppk-logo.svg';
 import unknownLogo from '!!@svgr!./unknown-logo.svg';
 import unknownNordicLogo from '!!@svgr!./unknown-nordic-logo.svg';
+
+type Device = NrfdlDevice | WrappedDevice;
 
 type DevicePCA =
     | 'PCA10028'
@@ -184,7 +188,7 @@ const deviceByPca = (device: Device) =>
 
 const NORDIC_VENDOR_ID = '1915';
 const isNordicDevice = (device: Device) =>
-    device.serialport?.vendorId === NORDIC_VENDOR_ID;
+    device.serialPorts?.[0]?.vendorId === NORDIC_VENDOR_ID;
 
 const ppkDeviceInfo = (device: Device): DeviceInfo => ({
     name: device.usb?.product,
@@ -217,11 +221,16 @@ const unknownDevice = (device: Device): DeviceInfo => ({
 export const deviceInfo = (device: Device): DeviceInfo =>
     deviceByPca(device) || deviceByUsb(device) || unknownDevice(device);
 
+const hasNickname = (
+    device: Device
+): device is WrappedDevice & Required<Pick<WrappedDevice, 'nickname'>> =>
+    'nickname' in device && device.nickname !== undefined;
+
 export const displayedDeviceName = (
     device: Device,
     { respectNickname = true } = {}
 ) => {
-    if (respectNickname && device.nickname) {
+    if (respectNickname && hasNickname(device)) {
         return device.nickname;
     }